dom/tests/mochitest/browser-frame/test_browserFrame7.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=725796
-->
<head>
  <title>Test for Bug 725796</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="browserFrameHelpers.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=725796">Mozilla Bug 725796</a>

<!--
  Test that an <iframe mozbrowser> is a window.{top,parent,frameElement} barrier.
-->

<script type="application/javascript;version=1.7">
"use strict";

browserFrameHelpers.setEnabledPref(true);
browserFrameHelpers.addToWhitelist();

// This test works only with OOP browser frames.  If browserFramesHelpers
// didn't decide to enable OOP browser frames, just skip this test.
if (browserFrameHelpers.getOOPDisabledPref()) {
  ok(true, "Skipping test because OOP browser frames are disabled.");
}
else {
  runTest();
}

var iframe;
function runTest() {
  iframe = document.createElement('iframe');
  iframe.addEventListener('mozbrowserloadend', function() {
    try {
      outerIframeLoaded();
    } catch(e) {
      dump("Got error: " + e + '\n');
    }
  });
  iframe.mozbrowser = true;
  iframe.src = 'data:text/html,Outer iframe <iframe id="inner-iframe"></iframe>';
  // For kicks, this test uses a display:none iframe.  This shouldn't make a
  // difference in anything.
  iframe.style.display = 'none';
  document.body.appendChild(iframe);

  SimpleTest.waitForExplicitFinish();
}

var numMsgReceived = 0;
function outerIframeLoaded() {
  var injectedScript = 
    "data:,function is(a, b, desc) {                                     \
      if (a == b) {                                                      \
        sendAsyncMessage('test:test-pass', desc);                        \
      } else {                                                           \
        sendAsyncMessage('test:test-fail', desc + ' ' + a + ' != ' + b); \
      }                                                                  \
    }                                                                    \
    is(content.window.top, content.window, 'top');                       \
    is(content.window.parent, content.window, 'parent');                 \
    is(content.window.frameElement, null, 'frameElement');               \
    var innerIframe = content.document.getElementById('inner-iframe');   \
    var innerWindow = innerIframe.contentWindow;                         \
    is(innerWindow.top, content.window, 'inner top');                    \
    is(innerWindow.parent, content.window, 'inner parent');              \
    is(innerWindow.frameElement, innerIframe, 'inner frameElement');"

  var mm = SpecialPowers.getBrowserFrameMessageManager(iframe);

  function onRecvTestPass(msg) {
    numMsgReceived++;
    ok(true, msg.json);
  }
  mm.addMessageListener('test:test-pass', onRecvTestPass); 

  function onRecvTestFail(msg) {
    numMsgReceived++;
    ok(false, msg.json);
  }
  mm.addMessageListener('test:test-fail', onRecvTestFail);

  mm.loadFrameScript(injectedScript, /* allowDelayedLoad = */ false);

  waitForMessages(6);
}

function waitForMessages(num) {
  if (numMsgReceived < num) {
    SimpleTest.executeSoon(function() { waitForMessages(num); });
    return;
  }
  
  SimpleTest.finish();
}

</script>
</body>
</html>
